backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12
aiofiles==24.1.0

# Database (Neon PostgreSQL)
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.2.5

# AI/LLM APIs
ollama==0.4.0
google-generativeai==0.8.0
groq==0.11.0

# Embeddings (for pgvector)
sentence-transformers==2.2.2

# Data Processing - UPDATED FOR PYTHON 3.12
pandas==2.2.3
numpy==1.26.4
PyPDF2==3.0.1

# HTTP & Networking
httpx==0.27.2
requests==2.31.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.0
pillow==11.0.0
tqdm==4.66.1